rework reader/writer/schema imports
