+++ fastavro/_writer_common.py
@@ -1,0 +1,12 @@
+_base_types = [
+    'boolean',
+    'bytes',
+    'double',
+    'float',
+    'int',
+    'long',
+    'null',
+    'string',
+]
+
+SCHEMA_DEFS = {typ: typ for typ in _base_types}
